update to 1.19

+++ src/main/java/net/maroonangel/keepchat/mixin/ChatScreenMixin.java
@@ -6,7 +6,6 @@ import net.minecraft.client.gui.screen.CommandSuggestor;
 import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.client.gui.widget.TextFieldWidget;
 import net.minecraft.text.Text;
-import net.minecraft.text.TranslatableText;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Overwrite;
 import org.spongepowered.asm.mixin.Shadow;
@@ -42,9 +41,8 @@ public abstract class ChatScreenMixin extends Screen {
             } else
                 amount *= 2.0D;
 
-            this.client.inGameHud.getChatHud().scroll(amount);
+            this.client.inGameHud.getChatHud().scroll((int)amount);
             return true;
         }
     }
-
 }
